requirements.txt
bcrypt==4.1.3
boto3==1.34.139
botocore==1.34.139
click==8.1.7
fastapi==0.111.0
fastapi-cli==0.0.5
gitdb==4.0.11
GitPython==3.1.43
Jinja2==3.1.4
langchain==0.2.8
numpy==1.26.4
passlib==1.7.4
pydantic==2.8.2
pymongo==4.8.0
python-dotenv==1.0.1
referencing==0.35.1
regex==2024.7.24
replicate==0.27.0
requests==2.32.3
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
starlette==0.37.2
uvicorn==0.30.1
websockets==12.0
yarl==1.9.4
gunicorn
python-jose==3.3.0
Pillow>=10.0.0
Wand==0.6.13
pydantic-settings==2.7.1